Co-locate related routes
