+++ backend/src/main.rs
@@ -69,12 +69,18 @@ async fn main() {
     let app = Router::new()
         .merge(SwaggerUi::new("/docs").url("/docs/openapi.json", ApiDoc::openapi()))
         .route("/", get(get_root))
-        .route("/stats", get(get_stats))
-        .route("/event", post(create_event))
-        .route("/event/:event_id", get(get_event))
-        .route("/event/:event_id/people", get(get_people))
-        .route("/event/:event_id/people/:person_name", get(get_person))
-        .route("/event/:event_id/people/:person_name", patch(update_person))
+        .route("/stats", get(stats::get_stats))
+        .route("/event", post(event::create_event))
+        .route("/event/:event_id", get(event::get_event))
+        .route("/event/:event_id/people", get(person::get_people))
+        .route(
+            "/event/:event_id/people/:person_name",
+            get(person::get_person),
+        )
+        .route(
+            "/event/:event_id/people/:person_name",
+            patch(person::update_person),
+        )
         .with_state(shared_state)
         .layer(cors)
         .layer(rate_limit)
