--- backend/src/routes/get_event.rs
@@ -1,42 +1,0 @@
-use axum::{
-    extract::{self, Path},
-    Json,
-};
-use common::adaptor::Adaptor;
-
-use crate::{
-    errors::ApiError,
-    payloads::{ApiResult, EventResponse},
-    State,
-};
-
-#[utoipa::path(
-    get,
-    path = "/event/{event_id}",
-    params(
-        ("event_id", description = "The ID of the event"),
-    ),
-    responses(
-        (status = 200, description = "Ok", body = EventResponse),
-        (status = 404, description = "Not found"),
-        (status = 429, description = "Too many requests"),
-    ),
-    tag = "event",
-)]
-/// Get details about an event
-pub async fn get_event<A: Adaptor>(
-    extract::State(state): State<A>,
-    Path(event_id): Path<String>,
-) -> ApiResult<EventResponse, A> {
-    let adaptor = &state.lock().await.adaptor;
-
-    let event = adaptor
-        .get_event(event_id)
-        .await
-        .map_err(ApiError::AdaptorError)?;
-
-    match event {
-        Some(event) => Ok(Json(event.into())),
-        None => Err(ApiError::NotFound),
-    }
-}
