+++ backend/src/routes/mod.rs
@@ -1,17 +1,3 @@
-pub mod get_event;
-pub use get_event::get_event;
-
-pub mod get_stats;
-pub use get_stats::get_stats;
-
-pub mod create_event;
-pub use create_event::create_event;
-
-pub mod get_people;
-pub use get_people::get_people;
-
-pub mod get_person;
-pub use get_person::get_person;
-
-pub mod update_person;
-pub use update_person::update_person;
+pub mod event;
+pub mod person;
+pub mod stats;
