+++ backend/src/routes/stats.rs
